Mark fbcode toolchains as rust_nightly
Summary:
Fbcode toolchains support nightly features. The allocative build
script checks `rustc --version` for the presence of the nightly substring, but
the fbcode toolchain won't have it in its output.

Add an additional check to see if the sysroot for rustc contains `platform010`,
indicating that the fbcode platform is used and nightly features can be
enabled.

Reviewed By: ndmitchell, krallin

Differential Revision: D41117401

fbshipit-source-id: c4fe06ce1d774afa49c821a558f97c93a817c330