Implement simple non-empty verbatim bound (#6)
Summary:
This implements the non-empty `allocative(bound = "...")` as a simple verbatim bounds override. This is similar to how `serde` handles its `bound` annotations where the user simply gets full control to add whatever they like. The "downside" of this is that the user *must* make sure to list all type variables that are used in the type.

The upshot is that it gives a lot of flexibility to the user to also list additional constraints,(e.g. lifetimes) and it is very efficient to imlpement.

For example

```
#[derive(allocative::Allocative)]
#[allocative(bound = "N: allocative::Allocative, E: allocative::Allocative, Ty")]
pub struct GraphMap<N, E, Ty> {
    nodes: IndexMap<N, Vec<(N, CompactDirection)>>,
    edges: IndexMap<(N, N), E>,
    #[allocative(skip)]
    ty: PhantomData<Ty>,
}
```

Creates

```
impl<N: allocative::Allocative, E: allocative::Allocative, Ty> allocative::Allocative for GraphMap<N, E, Ty> {
    ....
}
```

Pull Request resolved: https://github.com/facebookexperimental/allocative/pull/6

Reviewed By: ndmitchell

Differential Revision: D52629410

Pulled By: stepancheg

fbshipit-source-id: a115886ac65c8e1f5f13b2633e3c6893054e74e2